Fix broken pointer logic
Add pointer to the set of providers, not the querying node. This bug was introduced because the assumption in IPFS that [`p == pi.ID`](https://github.com/libp2p/go-libp2p-kad-dht/blob/master/handlers.go#L279) does not hold for pointers.

Not quite sure why the testbench didn't catch this. @jackkleeman and I have lost many hours on this.